ADD RxDocument().synced$
