+++ config/gulpfile.js
@@ -12,9 +12,12 @@ gulp.task('test', function() {
             '../test_tmp/node/RxDatabase.test.js',
             '../test_tmp/node/RxCollection.test.js',
             '../test_tmp/node/RxDocument.test.js',
-            '../test_tmp/node/Observe.test.js',
+            '../test_tmp/node/Reactive-Database.test.js',
+            '../test_tmp/node/Reactive-Collection.test.js',
+            '../test_tmp/node/Reactive-Query.test.js',
+            '../test_tmp/node/Reactive-Document.test.js',
             '../test_tmp/node/Primary.test.js',
-            '../test_tmp/node/Hooks.test.js',          
+            '../test_tmp/node/Hooks.test.js',
             '../test_tmp/node/ORM.test.js',
             '../test_tmp/node/Population.test.js',
             '../test_tmp/node/DataMigration.test.js',
